move TextInput
