+++ src/lib/components/common/TextInput/index.ts
@@ -1,0 +1,3 @@
+import TextInput from './TextInput.svelte';
+
+export default TextInput;
